+++ modules/angular2/src/core/compiler/pipeline/element_binder_builder.js
@@ -11,7 +11,7 @@ import {CompileStep} from './compile_step';
 import {CompileElement} from './compile_element';
 import {CompileControl} from './compile_control';
 import {dashCaseToCamelCase} from '../string_utils';
-import {setterFactory} from '../property_setter_factory'
+import {setterFactory} from 'angular2/src/render/dom/compiler/property_setter_factory'
 
 /**
  * Creates the ElementBinders and adds watches to the
