+++ modules/angular2/src/render/dom/compiler/selector.js
