+++ modules/angular2/src/render/dom/shadow_dom/shadow_css.js
