+++ modules/angular2/src/render/dom/shadow_dom/style_inliner.js
@@ -1,9 +1,10 @@
 import {Injectable} from 'angular2/di';
 import {XHR} from 'angular2/src/services/xhr';
-import {StyleUrlResolver} from 'angular2/src/core/compiler/style_url_resolver';
-import {UrlResolver} from 'angular2/src/services/url_resolver';
-
 import {ListWrapper} from 'angular2/src/facade/collection';
+
+import {UrlResolver} from 'angular2/src/services/url_resolver';
+import {StyleUrlResolver} from './style_url_resolver';
+
 import {
   isBlank,
   isPresent,
