+++ modules/angular2/src/render/dom/shadow_dom/style_url_resolver.js
