+++ modules/angular2/test/core/compiler/shadow_dom_strategy_spec.js
@@ -18,8 +18,8 @@ import {
   resetShadowDomCache,
 } from 'angular2/src/core/compiler/shadow_dom_strategy';
 import {UrlResolver} from 'angular2/src/services/url_resolver';
-import {StyleUrlResolver} from 'angular2/src/core/compiler/style_url_resolver';
-import {StyleInliner} from 'angular2/src/core/compiler/style_inliner';
+import {StyleUrlResolver} from 'angular2/src/render/dom/shadow_dom/style_url_resolver';
+import {StyleInliner} from 'angular2/src/render/dom/shadow_dom/style_inliner';
 import {ProtoView} from 'angular2/src/core/compiler/view';
 import {DirectiveMetadata} from 'angular2/src/core/compiler/directive_metadata';
