+++ modules/angular2/test/render/dom/compiler/selector_spec.js
@@ -1,7 +1,7 @@
 import {describe, it, expect, beforeEach, ddescribe, iit, xit, el} from 'angular2/test_lib';
 import {DOM} from 'angular2/src/dom/dom_adapter';
-import {SelectorMatcher} from 'angular2/src/core/compiler/selector';
-import {CssSelector} from 'angular2/src/core/compiler/selector';
+import {SelectorMatcher} from 'angular2/src/render/dom/compiler/selector';
+import {CssSelector} from 'angular2/src/render/dom/compiler/selector';
 import {List, ListWrapper, MapWrapper} from 'angular2/src/facade/collection';
 
 export function main() {
