+++ modules/benchmarks/src/compiler/selector_benchmark.js
@@ -1,5 +1,5 @@
-import {SelectorMatcher} from "angular2/src/core/compiler/selector";
-import {CssSelector} from "angular2/src/core/compiler/selector";
+import {SelectorMatcher} from "angular2/src/render/dom/compiler/selector";
+import {CssSelector} from "angular2/src/render/dom/compiler/selector";
 import {StringWrapper, Math} from 'angular2/src/facade/lang';
 import {ListWrapper} from 'angular2/src/facade/collection';
 import {getIntParameter, bindAction} from 'angular2/src/test_lib/benchmark_util';
